requirements.txt
apache-airflow==2.9.3
requests>=2.31.0
praw>=7.7.0
python-dotenv>=1.0.1
pre-commit>=3.8.0
black>=24.8.0
flake8>=7.1.1
google-cloud-storage>=2.17.0
pandas>=2.2.2
great_expectations>=0.18.5
python-snappy>=0.7.1
dbt-core>=1.8.0
dbt-bigquery>=1.8.0
google-cloud-bigquery>=3.17.0